# src/main/java/com/example/devices/dto/DeviceDto.java
package com.example.devices.dto;

import java.time.LocalDateTime;

public class DeviceDto {
    private Long id;
    private String name;
    private String brand;
    private String state;
    private LocalDateTime creationTime;

    public DeviceDto() {
    }

    public DeviceDto(Long id, String name, String brand, String state, LocalDateTime creationTime) {
        this.id = id;
        this.name = name;
        this.brand = brand;
        this.state = state;
        this.creationTime = creationTime;
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getBrand() {
        return brand;
    }

    public void setBrand(String brand) {
        this.brand = brand;
    }

    public String getState() {
        return state;
    }

    public void setState(String state) {
        this.state = state;
    }

    public LocalDateTime getCreationTime() {
        return creationTime;
    }

    public void setCreationTime(LocalDateTime creationTime) {
        this.creationTime = creationTime;
    }
}